refactor: Docker containerization support and config module reorganization
- Rename config_util.py -> app_config.py (application config: get_root_path, get_conf, load_configs)
- Rename conf_util.py -> ssl_config.py (SSL cert config: load_conf_obj, load_cert_password)
- Consolidate duplicated env override logic into app_config.apply_env_overrides()
- Move persistence config (get_persistence_conf) and shared utilities into app_config
- Add REGISTRY_* environment variable override mechanism for Docker deployments
- Add Dockerfile (Python 3.10-slim), docker-compose.yml, docker-entrypoint.sh
- Entrypoint auto-generates self-signed TLS + signing certificates on first run
- Add .dockerignore to exclude dev files from Docker builds
