+++ common/util/constant_param.py
@@ -15,7 +15,7 @@
 
 import os
 
-from common.util.config_util import get_root_path
+from common.util.app_config import get_root_path
 
 ROOT_PATH = get_root_path()
 SSL_PATH = os.path.join(ROOT_PATH, "etc", "ssl")
