+++ common/vector_db/__init__.py
@@ -12,7 +12,7 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
-from common.util.config_util import get_conf
+from common.util.app_config import get_conf
 
 conf = get_conf()
 use_vector_db = str(conf.get("use_vectordb", False)).lower() == 'true'
