+++ common/vector_db/vector_db_client/milvus_client.py
@@ -19,7 +19,7 @@ from pymilvus import MilvusClient, DataType, MilvusException
 from loguru import logger
 
 from agent_registry.config import AGENT_NUM_MAX
-from common.util.config_util import get_conf
+from common.util.app_config import get_conf
 from common.vector_db.vector_db_client.config.vector_db_client import VectorDBClient
 from common.vector_db.vector_db_client.config.vector_db_client_registry import vectordb_tool_register
 from common.vector_db.vector_db_client.config.vector_db_config import VectorDBType
